titled-dialog: Show min and max buttons (Fixes #26)
Since GtkDialogs with CSD only have a close button and our settings
dialogs are only "dialogs" we want to show the min and max buttons too
there. This means we have to set the type-hint for the window manager to
"normal" and to be sure we also manually center the window on the active
screen.