shortcuts-grabber: Redesign shortcut regrabbing (Bug #33)
This patch hopes to fix shortcut grabbing issues related to keyboard
layouts while maintaining high performance.

The implementation uses a new hash-table for tracking the keycodes grabbed
from the X server. The grabbed X11/Xorg keys are reference counted:
X11 XGrabKey() is called once per a keycode+modifiers combination and
X11 XUngrabKey() is called when the reference count of the combination
drops to zero. It is common for the reference counts to, for example,
reach the value of 4 if the user is using 4 keyboard layouts, in which
case the new implementation will use a single XGrabKey() call compared
to 4 such calls in previous implementations.

The grab_all() function has been removed and has been replaced by an
optimized regrab_all() function which is more efficient than the sequence
ungrab_all()+grab_all().

Tested keyboard layouts:
 - English
 - English (Colemak)
 - French (BÉPO)
 - Slovak (QWERTY)

Test environments:
 - Arch Linux (FR-BÉPO)
 - FreeBSD (basic testing)
 - Gentoo Linux (EN, EN-Colemak, SK-QWERTY)
 - Xubuntu (basic testing)

Closes: https://gitlab.xfce.org/xfce/libxfce4ui/-/issues/33