XfceShortcutsProvider: Add auto-repeat support
An extended setter is introduced instead of a simple separate setter for
"auto-repeat", because otherwise the "shortcut-added" signal is emitted
before the new property is set. From this point on, it is better to
introduce a setter with a variable number of arguments in case more
should be added in the future.

Related: !63, xfce4-settings#422
MR: !85