shortcuts: Fix Xfconf memory management
The code assumed that xfconf_init() was called by the main program,
which is not specified anywhere and is not obvious, although it is
probably still true in practice in the Xfce code base. Also,
xfconf_channel_new() was used instead of xfconf_channel_get(), which is
discouraged.

That said, assuming that xfconf_init() was properly called upstream,
there was no memory management error per se here. So this commit is more
like a cleanup.