shortcuts-grabber: Add setting to ignore keyboard layout change
b50af342 added support for keyboard layout change to fix some bugs,
however this appears to be a regression for some users. This commit adds
a preference to ignore layout change again, so that the old behavior is
restored.

In practice this should simply undo the effect of b50af342 and
subsequent commits in !27, always using the first available keyboard
layout. Enabling this option should therefore make the bugs fixed in !27
reappear: it seems difficult to satisfy everyone at once in this case.

Related: xfce4-settings#423, !27