screensaver: Prevent recursive calling of lock command
This is not intended to prevent a malicious attack but an accidental
recursive call. When xfce4-session!41 is merged, this will allow in
particular to address xfce4-session#120 in this sense. The original idea
is from Jarno Suni in xfce4-session!20.

(cherry picked from commit 1678d6082c5032a8bf9da591c7251ee8452e7250)