shortcuts-grabber: Remove filtering by level
This basically reverts b2e1dac2 and a0203505. The reason is that I don't
know how to tell the difference between legitimate and illegitimate
filtering. In particular, using consumed modifiers doesn't help. In the
end, it's better to grab a few too many keys than not enough, so the
balance tips in favor of deletion.

Closes: #101