build: Fix regenerating libxfce4ui.h after configure options change
Add a dependency on `config.h` to ensure that `libxfce4ui.h` is
regenerate whenever `./configure` options change.  Otherwise, once
generated the header is never invalidated, and the X11 variant could
be installed with `--disable-x11` or the other way around.

This is particularly a problem because X11-enabled header is included
in the source distribution, effectively preventing it from being updated
even if the package is initially configured with `--disable-x11`.

Originally reported as https://bugs.gentoo.org/924164.