Add "exclusive groups" to the keyboard shortcuts editor
The idea of an exclusive group is to allow some, but not all, shortcuts
to overlap.  For example, you may have a shortcuts editor that defines
shortcuts for two (or more) different types of application window, where
the different window types will not get the same list of shortcut
sections applied to them.  In that case, it is completely valid for
a section to have shortcuts that conflicts with shortcuts in another
section.

Shortcuts within a section cannot overlap, and the default behavior (if
no exclusive groups are added) will still be to assume that no shortcuts
in all sections can overlap.

Closes #115