Add "overlap groups" to the keyboard shortcuts editor
The idea of an overlap group is to allow some, but not all, shortcuts to overlap.  For example, you may have a shortcuts editor that defines
shortcuts for two (or more) different types of application window, where the different window types will not get the same list of shortcut sections applied to them.  In that case, it is completely valid for a section to have shortcuts that conflicts with shortcuts in another section.

Shortcuts within a section cannot overlap.

Closes #115