Add "overlap groups" to the shortcut editor (#115)
The idea of an overlap group is to allow some, but not all, shortcuts to overlap. 

For example, you may have a shortcut editor that defines shortcuts for two (or more) different types of application window, where the different window types will not get the same list of shortcut sections applied to them. In that case, it is completely valid for a section to have shortcuts that conflicts with shortcuts in another section.

Closes #115

Co-authored-by: default avatarBrian Tarricone <brian@tarricone.org>