Add variadic functions to Vala
GObject introspection does not currently support variadic methods. It
actually generates all the information Vala needs to do so, but it marks
the functions as introspectable="0", which is the same that happens when
"skip" annotation is added to the method. Therefore, in order to expose
these functions in Vala, we need a simple annotation to un-skip the
symbols.

References:
1. https://wiki.gnome.org/Projects/Vala/Bindings
2. https://gitlab.gnome.org/GNOME/vala/-/blob/main/vapi/metadata/Gtk-3.0.metadata